Rename / Transition to Headless UI Tab Control
