+++ frontend-next/src/app/app/page.js
@@ -10,7 +10,7 @@ import { useGeolocated } from "react-geolocated";
 import { Header } from "../../components/app/header";
 
 // Main Tab Imports
-import { MainTabHome } from "../../components/app/main_tab/home";
+import { MainTabHome } from "../../components/app/page/home";
 
 // Sidebar Imports
 import { Home_Sidebar } from "../../components/app/sidebar/home";
